requirements.txt
# Core Dependencies
PyPDF2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
groq>=0.4.0

# NLP & Processing
nltk>=3.8.0
spacy>=3.7.0
tabulate>=0.9.0

# Database
# SQLite is built-in; no extra package needed

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.24.0
